Fix SystemEnergy decode — use Int instead of UInt8/UInt16
Device JSON sends numeric values that JSONDecoder may not decode as
UInt8/UInt16 depending on the underlying NSNumber type. Using Int
is universally compatible with JSON number types.

+++ Sources/MovesenseApi/MovesenseTypes.swift
@@ -181,9 +181,9 @@ public struct MovesenseIMU: Codable {
 
 public struct MovesenseSystemEnergy: Codable {
 
-    public let percentage: UInt8
-    public let milliVolts: UInt16?
-    public let internalResistance: UInt8?
+    public let percentage: Int
+    public let milliVolts: Int?
+    public let internalResistance: Int?
 }
 
 public struct MovesenseSystemMode: Codable {
